arenda_site/requirements.txt
aiohttp==3.8.1
aioredis==1.3.1
aiosignal==1.2.0
amqp==5.0.9
asgiref==3.5.0
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
Babel==2.9.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.11
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.2
django-filter==21.1
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
Faker==12.0.1
flower==1.0.0
frozenlist==1.3.0
gunicorn==20.1.0
h11==0.13.0
hiredis==2.0.0
humanize==4.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.3
incremental==21.3.0
kombu==5.2.3
lml==0.1.0
lxml==4.7.1
Markdown==3.3.7
mixer==7.2.1
msgpack==1.0.3
multidict==6.0.2
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
phonenumbers==8.12.48
Pillow==9.0.1
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyexcel-io==0.6.6
pyexcel-xlsx==0.6.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
pyTelegramBotAPI==4.4.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-monkey-business==1.0.0
pytz==2021.3
redis==4.1.3
requests==2.27.1
selenium==4.1.5
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sorl-thumbnail==12.8.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
tornado==6.1
trio==0.20.0
trio-websocket==0.9.2
Twisted==22.1.0
txaio==21.2.1
typing-extensions==4.0.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.0.0
wrapt==1.13.3
wsproto==1.1.0
yarl==1.7.2
zipp==3.8.0
zope.interface==5.4.0
